Cleaned up docstrings, typos etc. Basically Done.
Still have a bug in randomDag. Just an idea: possibly we could have  a class MarkovChain in which an object is instantiated at the empty DAG and has methods which mutate it to the next element in the chain... maybe this way we could compute the elements in the chain more efficiently by storing data about the current element in the attributes.